pull dislike service to dislikes-service.js
