+++ src/services/dislikes-service.js
@@ -1,0 +1,17 @@
+import axios from "axios";
+
+const BASE_URL = process.env.REACT_APP_BASE_URL;
+const USERS_API = `${BASE_URL}/api/users`;
+const TUITS_API = `${BASE_URL}/api/tuits`;
+
+const api = axios.create({
+    withCredentials: true
+});
+
+export const userDislikesTuit = (uid, tid) =>
+    api.put(`${USERS_API}/${uid}/dislikes/${tid}`)
+        .then(response => response.data);
+
+export const tuitDislikedByUser = (uid, tid) =>
+    api.get(`${USERS_API}/${uid}/dislikes/${tid}`)
+        .then(response => response.data);
